requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
pymysql
cryptography
python-dotenv